+++ actionpack/lib/action_dispatch/middleware/cookies.rb
@@ -194,7 +194,12 @@ module ActionDispatch
   #     `:lax`.
   #
   class Cookies
-    HTTP_HEADER   = "Set-Cookie"
+    include ActiveSupport::Deprecation::DeprecatedConstantAccessor
+
+    deprecate_constant :HTTP_HEADER, "Rack::SET_COOKIE",
+      deprecator: ActionDispatch.deprecator,
+      message: "ActionDispatch::Cookies::HTTP_HEADER is deprecated and will be removed in Rails 9.0. Use Rack::SET_COOKIE instead."
+
     GENERATOR_KEY = "action_dispatch.key_generator"
     SIGNED_COOKIE_SALT = "action_dispatch.signed_cookie_salt"
     ENCRYPTED_COOKIE_SALT = "action_dispatch.encrypted_cookie_salt"